Yeouido_Parking_Flutter/python/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
PyMySQL==1.1.1
python-dotenv==1.1.0
email-validator==2.2.0
